+++ packages/chopsticks/src/plugins/index.ts
@@ -1,5 +1,5 @@
 import { Handlers } from '@acala-network/chopsticks-core'
-import { lstatSync, readdirSync } from 'fs'
+import { existsSync, lstatSync, readdirSync } from 'fs'
 import _ from 'lodash'
 import type { Argv } from 'yargs'
 
@@ -7,18 +7,23 @@ import { defaultLogger } from '../logger.js'
 
 const logger = defaultLogger.child({ name: 'plugin' })
 
-export const pluginHandlers: Handlers = {}
+export const rpcPluginHandlers: Handlers = {}
 
 // list of plugins directory
 const plugins = readdirSync(new URL('.', import.meta.url)).filter((file) =>
   lstatSync(new URL(file, import.meta.url)).isDirectory(),
 )
 
+// find all rpc exports
+export const rpcPluginMethods = readdirSync(new URL('.', import.meta.url))
+  .filter((file) => existsSync(new URL(`${file}/rpc.ts`, import.meta.url)))
+  .map((name) => `dev_${_.camelCase(name)}`)
+
 export const loadRpcPlugin = async (method: string) => {
   if (!process.env.DISABLE_PLUGINS) {
     return undefined
   }
-  if (pluginHandlers[method]) return pluginHandlers[method]
+  if (rpcPluginHandlers[method]) return rpcPluginHandlers[method]
 
   const plugin = _.snakeCase(method.split('dev_')[1]).replaceAll('_', '-')
   if (!plugin) return undefined
@@ -28,7 +33,7 @@ export const loadRpcPlugin = async (method: string) => {
   const { rpc } = await import(location.pathname)
   if (!rpc) return undefined
 
-  pluginHandlers[method] = rpc
+  rpcPluginHandlers[method] = rpc
   logger.debug(`Registered plugin ${plugin} RPC`)
 
   return rpc
