+++ packages/chopsticks/src/plugins/set-runtime-log-level/index.ts
@@ -1,27 +1,1 @@
-import { Context, ResponseError } from '@acala-network/chopsticks-core'
-import { defaultLogger } from '../../logger.js'
-
-/**
- * Set runtime log level.
- *
- * This function is a dev rpc handler. Use `dev_setRuntimeLogLevel` as the method name when calling it.
- *
- * @param context - The context object of the rpc handler
- * @param runtimeLogLevel - The runtime log level to set
- *
- * @example Set runtime log level to 1
- * ```ts
- * import { WsProvider } from '@polkadot/rpc-provider'
- * const ws = new WsProvider(`ws://localhost:8000`)
- * await ws.send('dev_setRuntimeLogLevel', [1])
- * ```
- */
-export const rpc = async (context: Context, [runtimeLogLevel]: [number]) => {
-  defaultLogger.debug({ runtimeLogLevel }, 'dev_setRuntimeLogLevel')
-
-  if (typeof runtimeLogLevel !== 'number') {
-    throw new ResponseError(1, `Invalid runtimeLogLevel ${runtimeLogLevel}`)
-  }
-
-  context.chain.runtimeLogLevel = runtimeLogLevel
-}
+export * from './rpc.js'
